rename include/nc-runtime to include/nc_runtime
