+++ include/nc_runtime/nc_log.h
@@ -32,9 +32,13 @@ void Dal_writeLog(const char* str);  // (DAL = device adaption layer) device dep
 #define NC_LOG_FATAL(...) __LOG_IMPLE(LogLevel::fatal, __FILE__, __LINE__, __FUNCTION__, __VA_ARGS__)
 
 #if defined(NDEBUG)
-#	define NC_ASSERT(predicate)
+#  define NC_ASSERT(predicate)
 #else
-#	define NC_ASSERT(predicate) if (!(predicate)) { NcLog_write(LogLevel::fatal, __FILE__, __LINE__, __FUNCTION__, "error: %s", #predicate); __builtin_trap(); }
+#  if defined(NC_OS_UNIX)
+#      define NC_ASSERT(predicate) if (!(predicate)) { NcLog_write(LogLevel::fatal, __FILE__, __LINE__, __FUNCTION__, "error: %s", #predicate); __builtin_trap(); }
+#  else
+#      define NC_ASSERT(predicate) if (!(predicate)) { NcLog_write(LogLevel::fatal, __FILE__, __LINE__, __FUNCTION__, "error: %s", #predicate); __debugbreak(); }
+#  endif
 #endif
 
 // clang-format on
