+++ src/nc_array.cpp
@@ -1,2 +1,2 @@
 #include "stdafx_nc_runtime.h"
-#include "nc_array.h"
+#include "nc_runtime/nc_array.h"
