+++ src/nc_string.cpp
@@ -1,6 +1,6 @@
 #include "stdafx_nc_runtime.h"
-#include "nc_string.h"
-#include "nc_stdlib.h"
+#include "nc_runtime/nc_string.h"
+#include "nc_runtime/nc_stdlib.h"
 #include <mutex>
 #include <unordered_map>
 #include <stdarg.h>
