+++ src/stdafx_nc_runtime.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "nc_types.h"
-#include "nc_stdlib.h"
+#include "nc_runtime/nc_types.h"
+#include "nc_runtime/nc_stdlib.h"
 
 using namespace std;
