+++ test/stdafx_nc_runtime_test.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "nc_types.h"
+#include "nc_runtime/nc_types.h"
 #include "gtest/gtest.h"
 
 using namespace std;
